statementSelect: Persistence lock timeout exception is caught.
When calling the treeStructureChanged (when there is a problem on live
inserting/deleting nodes on the tree) all the nodes affected must be
cleaned out from the nodeMap.

A cleanPath method to explictly updating the descending structure of a
node. That part of the tree will be "cleaned" and freed from memory.