Alternative text translations of a imported statement into the CliJPanel (drag & drop).
When the imported statement is in the path of the active context, the
relative path would be the empty string, which obviously is an ambiguous
specification. We use the identifier of the statement if exists and not
shaded by another statement down in the path, or the UUID if is not the
case.