Basically replicated the same modification as in commit 5d46c9963b7c162a2227c5988ac689986406ed10 for the re-checking of the signed proof status mark. So:
Modified the re-check of the signed proof status when deleting to try to
avoid the propagation of the algorithm to huge amounts of statements of
the database. 

When a deleted/turned to not signed proof statement provides a solution
for lots of context's consequents (>=100) we first check if there is any
other statement providing the same term (so it solves all that contexts
too) which we can ensure it will keep proven after the re-check. If so,
we can safely ignore all that contexts in the status reset routine.